High speed by buffer sequence on AVTs. Number of buffers specification on preference file. Don't auto create preference file if not there.

+++ labcams/utils.py
@@ -68,7 +68,7 @@ DEFAULTS = dict(cams = [{'description':'facecam',
 defaultPreferences = DEFAULTS
 
 
-def getPreferences(preffile = None):
+def getPreferences(preffile = None,create = False):
     ''' Reads the parameters from the home directory.
 
     pref = getPreferences(expname)
@@ -79,13 +79,15 @@ def getPreferences(preffile = None):
     if preffile is None:
         preffile = pjoin(preferencepath,'default.json')
 
-    if not os.path.isfile(preffile):
-        print('Creating preference file from defaults.')
+    if not os.path.isfile(preffile) and create:
+        display('Creating preference file from defaults.')
         if not os.path.isdir(preferencepath):
             os.makedirs(preferencepath)
         with open(preffile, 'w') as outfile:
             json.dump(defaultPreferences, outfile, sort_keys = True, indent = 4)
-            print('Saving default preferences to: ' + preffile)
+            display('Saving default preferences to: ' + preffile)
+    else:
+        display('File {0} does not exist.'.format(preffile))
     if os.path.isfile(preffile):
         if not os.path.isdir(preferencepath):
             os.makedirs(preferencepath)
